// frontend/src/app/models/user.model.ts
export interface User {
  id?: number;
  name: string;
  username: string;
  email: string;
  password?: string;
  role: 'USER' | 'CREATOR' | 'BUSINESS';
  bio?: string;
  location?: string;
  website?: string;
  profilePicturePath?: string;
  accountPrivacy: 'PUBLIC' | 'PRIVATE';
  createdAt?: string;
}

export interface LoginRequest {
  emailOrUsername: string;
  password: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}